Base64 encoding of images is the way to go! It increases network bandwidth by 1/3 - which is offset by me now using OpenCV to export the JPEG directly, with 50% compression. It's more work for the browser and server (compress/decompress), but the framerate is so low that it's quite manageable.
MJPEGs are just too much hassle. I created a standalone MJPEG streaming demo browsers have trouble knowing when they disconnect, and webkit has an open bug to this day where it won't close MJPeg streams until you navigate to a different page. When I got to the point of embedding them in iFrames to facilitate disconnecting without leaving the page, having the iframes dynamically resize using javascript, and using the websocket itself to communicate between the client and server when to change some property of the stream, I knew it was time to stop.

Everything's finally coming together... gotta code fast for tomorrow!